segregate input parsing and listener registration from serial event listener
