+++ src/main/java/org/chabala/brick/controllab/SerialPortEventListener.java
@@ -19,7 +19,6 @@
 package org.chabala.brick.controllab;
 
 import jssc.SerialPortEvent;
-import org.chabala.brick.controllab.sensor.SensorListener;
 
 /**
  * SerialPortEventListener.
@@ -40,12 +39,4 @@ interface SerialPortEventListener extends jssc.SerialPortEventListener {
     void serialEventRXCHAR(int availableBytes);
 
     boolean isHandshakeSeen();
-
-    void addStopButtonListener(StopButtonListener listener);
-
-    void removeStopButtonListener(StopButtonListener listener);
-
-    void addSensorListener(Input input, SensorListener listener);
-
-    void removeSensorListener(Input input, SensorListener listener);
 }
