+++ src/test/java/org/chabala/brick/controllab/JsscSerialPortTest.java
@@ -18,7 +18,6 @@
  */
 package org.chabala.brick.controllab;
 
-import jssc.SerialPortEventListener;
 import jssc.SerialPortException;
 import org.junit.Before;
 import org.junit.Rule;
